Fixed minor issues in the include files (#1290)

+++ src/mscrypto/globals.h
@@ -45,7 +45,7 @@
 
 void xmlSecMSCryptoGetErrorMessage      (DWORD dwError,
                                          xmlChar *out,
-                                         int outLen);
+                                         size_t outLen);
 
 
 /**
